seeded restaurant table

+++ database/migrations/2023_09_14_091921_create_restaurants_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->text("image");
             $table->string("address");
             $table->time("opening_time");
-            $table->smallInteger("P_IVA");
+            $table->integer("P_IVA");
             $table->timestamps();
         });
     }
